requirements.txt
# Requirements automatically generated by pigar.
# https://github.com/Damnever/pigar




# app.py: 6
requests == 2.10.0


click==6.7
Flask==0.12
Cython==0.25.2
funcsigs==1.0.2
h5py==2.6.0
itsdangerous==0.24
Jinja2==2.9.4s
Keras==1.2.1
MarkupSafe==0.23
mock==2.0.0
numpy==1.12.0
olefile==0.44
pbr==1.10.0
Pillow==4.0.0
protobuf==3.1.0.post1
PyYAML==3.12
scipy==0.18.1
six==1.10.0
tensorflow==0.12.1
Theano==0.8.2
Werkzeug==0.11.15
gunicorn==19.6.0